+++ src/game_engine/sys/gm_title_screen_class.cpp
@@ -37,7 +37,7 @@ void GmTitleScreen::init(ecs::Engine* ecs_engine)
 {
 	did_init = true;
 
-	size_t i = 0;
+	uint i = 0;
 	engine->aux_menu = Menu
 	(
 		"file_select",
